Do some cleanup as setup for the addr_t change.

+++ src/document.c
@@ -40,7 +40,7 @@ control_code_t check_for_control_code(uint8_t a)
     return NO_CONTROL_CODE;
 }
 
-int compute_bytes_free(void)
+ptrdiff_t compute_bytes_free(void)
 {
     // compute_bytes_free
     // Pseudocode: Computes number of free bytes between top and himem
@@ -57,7 +57,7 @@ int compute_bytes_free(void)
     //     tay
     // return_84:
     //     rts
-    return (int)himem - (int)top;
+    return himem - top;
 }
 
 void check_for_at_least_150_bytes_free(void)
